# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gin swagger 示例项目 首页展示",
        "title": "Gin swagger  展示在web端的title上",
        "contact": {},
        "license": {},
        "version": "1.0 定义接口的版本"
    },
    "paths": {
        "/index/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "主页接口",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/login/": {
            "post": {
                "description": "登录接口",
                "produces": [
                    "application/json"
                ],
                "summary": "登录接口",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.LoginForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "form.LoginForm": {
            "type": "object",
            "required": [
                "name",
                "password"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "description": "这里有做简单验证，表示参数是必须的",
                    "type": "string"
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth  API的认证方式": {
            "type": "apiKey",
            "name": "token  后端获取认证值得方式",
            "in": "cookie 发送认证的方式"
        }
    }
}